fix itemlot,itemsite
git-tfs-id: [https://hqsofttfs.hqsoft.com.vn/tfs/HQSOFT]$/Cloud/eSky4D;C31848

+++ PO/PO10201/PO10201/Controllers/PO10201Controller.cs
@@ -1,3 +1,4 @@
+ 
 using HQ.eSkyFramework;
 using Ext.Net;
 using Ext.Net.MVC;
@@ -40,8 +41,8 @@ namespace PO10201.Controllers
         string _handle = "";
         string _status = "";
         private JsonResult _logMessage;
-
-
+        private List<IN_ItemSite> lstInItemsiteNew = new List<IN_ItemSite>();
+        bool b235 = false;//message235
 
 
         public ActionResult Index()
@@ -51,7 +52,7 @@ namespace PO10201.Controllers
             ViewBag.BussinessTime = DateTime.Now;
             return View();
         }
-        [OutputCache(Duration = 1000000, VaryByParam = "lang")]
+        //[OutputCache(Duration = 1000000, VaryByParam = "lang")]
         public PartialViewResult Body(string lang)
         {
             return PartialView();
@@ -117,7 +118,7 @@ namespace PO10201.Controllers
                 _branchID = data["cboBranchID"];
                 _status = data["Status"].PassNull();
                 _handle = data["Handle"].PassNull() == "" ? _status : data["Handle"].PassNull();
-
+                b235 = _form["b235"].ToBool();
                 _objPO_Setup = _db.PO10201_pdPO_Setup(_branchID, "PO").FirstOrDefault();
 
                 var detHeader = new StoreDataHandler(data["lstHeader"]);
@@ -132,20 +133,14 @@ namespace PO10201.Controllers
                 _lstLot = detHandlerLot.ObjectData<PO10201_pgLotTrans_Result>()
                             .Where(p => Util.PassNull(p.LotSerNbr) != string.Empty)
                             .ToList();
-
+              
                 
-                if (Data_Checking())
+                if (Data_Checking(b235))
                 {
-                    if ((_status == "U" || _status == "C" || _status == "H") && (_handle == "C" || _handle == "V" || _handle == "R"))
+                    if ((_status == "U" || _status == "C"  ) && (_handle == "C" || _handle == "V"))
                     {
 
-                        if (_handle == "R" || _handle == "V" || _handle == "C")
-                        {
-                            if (_handle == "R" && !acc.Release)
-                            {
-                                throw new MessageException(MessageType.Message, "737");
-                            }
-                            else if ((_handle == "V" || _handle == "C") && !acc.Release)
+                       if ((_handle == "V" || _handle == "C") && !acc.Release)
                             {
                                 throw new MessageException(MessageType.Message, "725");
                             }
@@ -155,11 +150,17 @@ namespace PO10201.Controllers
                                 {
                                     Data_Release();
                                 }
-                                else if (_handle == "R") Save_Batch();
-                            }
-                        }
+                              
+                            }                      
                     }
-                    else Save_Batch();
+                    else if (_status == "H")
+                    {
+                        if (_handle == "R" && !acc.Release)
+                        {
+                            throw new MessageException(MessageType.Message, "737");
+                        }
+                        else Save_Batch();
+                    }
                 }
                 if (_logMessage != null)
                 {
@@ -231,6 +232,24 @@ namespace PO10201.Controllers
                             objItemSite.LUpd_DateTime = DateTime.Now;
                             objItemSite.LUpd_Prog = ScreenNbr;
                             objItemSite.LUpd_User = Current.UserName;
+
+                            // delete lot
+                            var lstold = _db.PO_LotTrans.Where(p => p.BranchID == obj.BranchID && p.BatNbr == obj.BatNbr && p.RefNbr == obj.RcptNbr && p.POTranLineRef == obj.LineRef).ToList();
+                            foreach (var objlot in lstold)
+                            {
+                                _db.PO_LotTrans.DeleteObject(objlot);
+                                double NewQty = (objlot.UnitMultDiv == "D" ? (objlot.Qty / objlot.CnvFact) : (objlot.Qty * obj.CnvFact));
+                                var objItemLot = _db.IN_ItemLot.Where(p => p.InvtID == objlot.InvtID && p.SiteID == obj.SiteID && p.LotSerNbr == objlot.LotSerNbr).FirstOrDefault();
+
+                                objItemLot.QtyAllocPORet = Math.Round(objItemLot.QtyAllocPORet - NewQty, 0);
+                                objItemLot.QtyAvail = Math.Round(objItemLot.QtyAvail + NewQty, 0);
+
+                                objItemLot.LUpd_DateTime = DateTime.Now;
+                                objItemLot.LUpd_Prog = ScreenNbr;
+                                objItemLot.LUpd_User = Current.UserName;
+
+
+                            }
                         }
                         _db.PO_Trans.DeleteObject(obj);
                         lstdel.Remove(obj);
@@ -294,7 +313,7 @@ namespace PO10201.Controllers
                             objItemSite.QtyAvail = Math.Round(objItemSite.QtyAvail + dblQty, 0);
                             objItemSite.LUpd_DateTime = DateTime.Now;
                             objItemSite.LUpd_Prog = ScreenNbr;
-                            objItemSite.LUpd_User = Current.UserName;
+                            objItemSite.LUpd_User = Current.UserName;                           
                         }
                         _db.PO_Trans.DeleteObject(obj);
                     }
@@ -333,6 +352,24 @@ namespace PO10201.Controllers
             return this.Direct(objIN_ItemSite);
 
         }
+        [DirectMethod]
+        public ActionResult PO10201ItemSiteQty(string branchID = "", string invtID = "", string siteID = "", string batNbr = "", string rcptNbr = "", string lineRef = "")
+        {
+            var objold = _db.PO_Trans.Where(p => p.BranchID == branchID && p.BatNbr == batNbr && p.RcptNbr == rcptNbr && p.InvtID == invtID && p.SiteID == siteID && p.LineRef == lineRef).FirstOrDefault();
+            var qtyold = objold == null ? 0 : objold.UnitMultDiv == "M" ? objold.Qty * objold.CnvFact : objold.Qty / objold.CnvFact;
+
+            var objIN_ItemSite = _db.IN_ItemSite.Where(p => p.InvtID == invtID && p.SiteID == siteID).FirstOrDefault();
+            if (objIN_ItemSite == null)
+            {
+                objIN_ItemSite = new IN_ItemSite();
+                objIN_ItemSite.ResetET();
+            }
+            objIN_ItemSite.QtyAvail=objIN_ItemSite.QtyAvail + qtyold;
+            return this.Direct(objIN_ItemSite);
+
+        }
+       
+
         [DirectMethod]
         public ActionResult INNumberingLot(string invtID = "", DateTime? tranDate = null, string getType = "LotNbr")
         {
@@ -470,30 +507,75 @@ namespace PO10201.Controllers
                     obj.tstamp = new byte[0];
                     _db.PO_Trans.AddObject(obj);
                 }
-               
+             
             }
-             Save_PO_LotTrans();
+            Save_PO_LotTrans();
         
         }
         private void Save_PO_LotTrans()
         {
             try
             {
-                // delete lot cu
+                //// delete lot cu khong co tren luoi lot
                 var lstold = _db.PO_LotTrans.Where(p => p.BranchID == _branchID && p.BatNbr == _batNbr && p.RefNbr == _rcptNbr).ToList();
                 foreach (var obj in lstold)
                 {
-                    _db.PO_LotTrans.DeleteObject(obj);
+                    if (_lstLot.Where(p => p.InvtID == obj.InvtID && p.SiteID == obj.SiteID && p.LotSerNbr == obj.LotSerNbr && p.POTranLineRef == obj.POTranLineRef).FirstOrDefault() == null)
+                    {
+                        _db.PO_LotTrans.DeleteObject(obj);
+                        if (_poHead.RcptType == "X")
+                        {
+                            double NewQty = (obj.UnitMultDiv == "D" ? (obj.Qty / obj.CnvFact) : (obj.Qty * obj.CnvFact));
+                            var objItemLot = _db.IN_ItemLot.Where(p => p.InvtID == obj.InvtID && p.SiteID == obj.SiteID && p.LotSerNbr == obj.LotSerNbr).FirstOrDefault();
+
+                            objItemLot.QtyAllocPORet = Math.Round(objItemLot.QtyAllocPORet - NewQty, 0);
+                            objItemLot.QtyAvail = Math.Round(objItemLot.QtyAvail + NewQty, 0);
+
+                            objItemLot.LUpd_DateTime = DateTime.Now;
+                            objItemLot.LUpd_Prog = ScreenNbr;
+                            objItemLot.LUpd_User = Current.UserName;
+                        }
+                    }
                 }
+              
 
                 //Save Lot/Serial from datatable to in_lottrans
 
                 foreach (var row in _lstLot)
                 {
-                    var obj = new PO_LotTrans();
-                    obj.ResetET();
-                    Update_PO_LotTrans(row, obj);
-                    _db.PO_LotTrans.AddObject(obj);
+                    double oldQty = 0;
+                    var obj = lstold.Where(p => p.BranchID == _branchID && p.BatNbr == _batNbr && p.RefNbr == _rcptNbr && p.InvtID == row.InvtID && p.LotSerNbr == row.LotSerNbr && p.SiteID == row.SiteID).FirstOrDefault();
+                    if (obj == null)
+                    {
+                        obj = new PO_LotTrans();
+                        obj.ResetET();
+                        Update_PO_LotTrans(row, obj);
+                        obj.Crtd_Prog = ScreenNbr;
+                        obj.Crtd_User = Current.UserName;
+                        obj.Crtd_DateTime = DateTime.Now;
+                        _db.PO_LotTrans.AddObject(obj);
+                    }
+                    else
+                    {
+                        oldQty = obj == null ? 0 : obj.UnitMultDiv == "M" ? obj.Qty * obj.CnvFact : obj.Qty / obj.CnvFact;
+                        Update_PO_LotTrans(row, obj);
+                    }
+                     //Update Location and Site Qty
+                    if (_poHead.RcptType == "X")
+                    {
+                        
+                        var qty = obj.UnitMultDiv == "M" ? obj.Qty * obj.CnvFact : obj.Qty / obj.CnvFact;                      
+                        var objItemLot = _db.IN_ItemLot.Where(p => p.InvtID == obj.InvtID && p.SiteID == obj.SiteID && p.LotSerNbr == obj.LotSerNbr).FirstOrDefault();
+
+                        objItemLot.QtyAllocPORet = Math.Round(objItemLot.QtyAllocPORet + qty - oldQty, 0);
+                        objItemLot.QtyAvail = Math.Round(objItemLot.QtyAvail - qty + oldQty, 0);
+
+                        objItemLot.LUpd_DateTime = DateTime.Now;
+                        objItemLot.LUpd_Prog = ScreenNbr;
+                        objItemLot.LUpd_User = Current.UserName;
+                        if (objItemLot.QtyAvail < 0)
+                            throw new MessageException(MessageType.Message, "35");
+                    }
                 }
                 _db.SaveChanges();
                 if (_handle == "R")
@@ -522,8 +604,8 @@ namespace PO10201.Controllers
                 objBatch.Module = "IN";
                 objBatch.JrnlType = "PO";
                 objBatch.Rlsed = 0;
-                objBatch.Status =(_handle != "N" && _handle!="")? _handle : _poHead.Status;
-
+                objBatch.Status = (_handle != "N" && _handle != "") ? _handle : _poHead.Status;
+               
                 objBatch.LUpd_DateTime = DateTime.Now;
                 objBatch.LUpd_Prog = ScreenNbr;
                 objBatch.LUpd_User = Current.UserName;
@@ -574,6 +656,7 @@ namespace PO10201.Controllers
                 objR.Status = (_handle != "N" && _handle != "") ? _handle : _poHead.Status;
 
 
+
                 objR.LUpd_DateTime = DateTime.Now;
                 objR.LUpd_Prog = ScreenNbr;
                 objR.LUpd_User = Current.UserName;
@@ -621,7 +704,7 @@ namespace PO10201.Controllers
                     var objIN_Inventory = _db.PO10201_pdIN_Inventory(Current.UserName).Where(p => p.InvtID == objr.InvtID).FirstOrDefault();
                     var objIN_ItemSite = _db.IN_ItemSite.Where(p => p.InvtID == objr.InvtID && p.SiteID == objr.SiteID).FirstOrDefault();
                     //Kiem tra itemsite neu chua co thi add vao
-                    if (objIN_ItemSite == null)
+                    if (objIN_ItemSite == null && lstInItemsiteNew.Where(p => p.InvtID == objr.InvtID && p.SiteID == objr.SiteID).Count()==0)
                     {
                         Insert_IN_ItemSite(ref objIN_ItemSite, ref objIN_Inventory, objr.SiteID);
                     }
@@ -778,9 +861,7 @@ namespace PO10201.Controllers
             objPO_LotTrans.UnitMultDiv = row.UnitMultDiv;
             objPO_LotTrans.UnitDesc = row.UnitDesc;
 
-            objPO_LotTrans.Crtd_Prog = ScreenNbr;
-            objPO_LotTrans.Crtd_User = Current.UserName;
-            objPO_LotTrans.Crtd_DateTime = DateTime.Now;
+          
             objPO_LotTrans.LUpd_Prog = ScreenNbr;
             objPO_LotTrans.LUpd_User = Current.UserName;
             objPO_LotTrans.LUpd_DateTime = DateTime.Now;
@@ -823,7 +904,7 @@ namespace PO10201.Controllers
                 objIN_ItemSite.tstamp = new byte[0];
 
                 _db.IN_ItemSite.AddObject(objIN_ItemSite);
-
+                lstInItemsiteNew.Add(objIN_ItemSite);
 
             }
             catch (Exception ex)
@@ -842,7 +923,7 @@ namespace PO10201.Controllers
         //    {
         //    }
         //}
-        private bool Data_Checking(bool isDeleteGrd = false)
+        private bool Data_Checking(bool isCheckInvoicePass = false)
         {
             if (_poHead.Status == "H")
             {
@@ -894,10 +975,10 @@ namespace PO10201.Controllers
                     throw new MessageException(MessageType.Message, "15", parm: new[] { Util.GetLang("DocDate") });
                 }
 
-                //if (_poHead.InvcDate.ToString().PassNull() == "")
-                //{
-                //    throw new MessageException(MessageType.Message, "15", parm: new[] { Util.GetLang("InvcDate") });
-                //}
+                if (_poHead.InvcDate.ToString().PassNull() == "")
+                {
+                    throw new MessageException(MessageType.Message, "15", parm: new[] { Util.GetLang("InvcDate") });
+                }
 
                 //Check PO has no detail data
                 if (_lstPOTrans.Count == 0)
@@ -921,29 +1002,53 @@ namespace PO10201.Controllers
                         throw new MessageException(MessageType.Message, "222");
 
                     }
-                    if (objPO_Trans.RcptQty == 0 || objPO_Trans.TranAmt == 0)
+                    if ((objPO_Trans.RcptQty == 0 || objPO_Trans.TranAmt == 0) && objPO_Trans.PurchaseType != "PR")
                     {
                         throw new MessageException(MessageType.Message, "44");
 
                     }
+
                 }
 
-                //kiem tra ton kho co du tra hang ko
+
                 if (_poHead.RcptType == "X")
                 {
-                    string invtID="";
+                    string invtID = "";
+
+                    //kiểm tra trong itemSite
+                    foreach (var objTran in _lstPOTrans)
+                    {
+                        var objold = _db.PO_Trans.Where(p => p.BranchID == _branchID && p.BatNbr == _batNbr && p.RcptNbr == _rcptNbr && p.InvtID == objTran.InvtID && p.SiteID == objTran.SiteID && p.LineRef == objTran.LineRef).FirstOrDefault();
+
+                        var qtyold = objold == null ? 0 : objold.UnitMultDiv == "M" ? objold.Qty * objold.CnvFact : objold.Qty / objold.CnvFact;
+                        var qty = objTran.UnitMultDiv == "M" ? objTran.Qty * objTran.CnvFact : objTran.Qty / objTran.CnvFact;
+                        var objItemSite = _db.IN_ItemSite.Where(p => p.InvtID == objTran.InvtID && p.SiteID == objTran.SiteID).FirstOrDefault();
+
+                        if (objItemSite == null || (qty - qtyold) > objItemSite.QtyAvail)
+                        {
+                            invtID += objTran.InvtID + ",";
+                        }
+
+                    }
+                    if (invtID != "") throw new MessageException(MessageType.Message, "1043", parm: new[] { invtID, "" });
+
                     //kiểm tra trong itemlot
                     foreach (var objlot in _lstLot)
                     {
+                        var objold = _db.PO_LotTrans.Where(p => p.BranchID == _branchID && p.BatNbr == _batNbr && p.RefNbr == _rcptNbr && p.POTranLineRef == objlot.POTranLineRef && p.LotSerNbr == objlot.LotSerNbr && p.InvtID == objlot.InvtID && p.SiteID == objlot.SiteID).FirstOrDefault();
+
+                        var qtyold = objold == null ? 0 : objold.UnitMultDiv == "M" ? objold.Qty * objold.CnvFact : objold.Qty / objold.CnvFact;
                         var qty = objlot.UnitMultDiv == "M" ? objlot.Qty * objlot.CnvFact : objlot.Qty / objlot.CnvFact;
                         var objItemLot = _db.IN_ItemLot.Where(p => p.InvtID == objlot.InvtID && p.SiteID == objlot.SiteID && p.LotSerNbr == objlot.LotSerNbr).FirstOrDefault();
-                        if (objItemLot==null || qty > objItemLot.QtyAvail)
+
+                        if (objItemLot == null || (qty - qtyold) > objItemLot.QtyAvail)
                         {
                             invtID += objlot.InvtID + ",";
                         }
-                        
+
                     }
-                    if (invtID != "") throw new MessageException(MessageType.Message, "1043",parm: new[] { invtID,""});
+                    if (invtID != "") throw new MessageException(MessageType.Message, "1043", parm: new[] { invtID, "" });
+
                 }
 
 
@@ -959,9 +1064,9 @@ namespace PO10201.Controllers
                 //}
 
                 //var obj1 = _db.PO10201_ppCheckExistingInvcNbr(_branchID, _batNbr, _poHead.VendID, _poHead.InvcNote, _poHead.InvcNbr).FirstOrDefault();
-                //if (obj1 != null)
+                //if (obj1 != null && !isCheckInvoicePass)
                 //{
-                //    throw new MessageException(MessageType.Message, "235");
+                //    throw new MessageException(MessageType.Message, "235", fn: "process235");
                 //}
 
 
@@ -972,7 +1077,7 @@ namespace PO10201.Controllers
 
         private void Data_Release()
         {
-            //if (_handle != "N" && _handle!="")
+            //if (_handle != "N")
             //{
             //    DataAccess dal = Util.Dal();
             //    try
